make options more stateful (#3)
Made the option type more useful, and replace the OptionHandler interface with Option. Connections start with an optionMap that is filled with regular options that implement the basic negotation, and basically ignore Update and Subnegotiation. Options can be added by embedding an Option in a new struct that adds more behavior or implements Update and Subnegotiation. Options are now bound to a Conn, and so have the Conn available to send things, which greatly simplifies all the option negotiation signatures.